refactor: 기타 refactoring 및 log.info 제거
