+++ src/main/java/com/example/outsourcingproject/config/AuthUserArgumentResolver.java
@@ -2,7 +2,7 @@ package com.example.outsourcingproject.config;
 
 import com.example.outsourcingproject.common.annotation.Auth;
 import com.example.outsourcingproject.common.dto.AuthUser;
-import com.example.outsourcingproject.domain.user.entity.UserRole;
+import com.example.outsourcingproject.domain.user.enums.UserRole;
 import jakarta.annotation.Nullable;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
