+++ src/main/java/com/example/outsourcingproject/domain/user/entity/User.java
@@ -1,6 +1,7 @@
 package com.example.outsourcingproject.domain.user.entity;
 
 import com.example.outsourcingproject.common.entity.BaseEntity;
+import com.example.outsourcingproject.domain.user.enums.UserRole;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
